requirements.txt
# The Witness / Jarvis - Reachy Mini Hackathon Project
# Install with: pip install -r requirements.txt
# Note: llama-cpp-python needs special build flags for CUDA (see README)

# Core ML
torch>=2.0
torchaudio
torchvision

# LLM inference (GGUF)
# Install separately with CUDA:
# CUDACXX=/usr/local/cuda/bin/nvcc CMAKE_ARGS="-DGGML_CUDA=on" pip install llama-cpp-python
llama-cpp-python

# Cloud LLM + ASR
openai
requests

# ASR
faster-whisper

# VLM
transformers
accelerate

# TTS
piper-tts

# Audio I/O
sounddevice

# Vision
opencv-python

# Hand pose estimation (install trt_pose + trt_pose_hand from git, see Dockerfile)
# torch2trt
# trt_pose
# trt_pose_hand
scikit-learn

# Model downloads
huggingface-hub

# Agora Conversational AI
agora-python-server-sdk==2.4.3
fastapi
uvicorn[standard]
httpx
mcp

# Utilities
numpy
